+++ assets/appFramework/react/src/typescript/child/child.tsx
@@ -2,14 +2,11 @@
  * Child class.
  */
 import * as React from "react";
-
-const childStyle = {
-    fontSize: "20px"
-};
+import "./child.css";
 
 export class Child extends React.Component {
     public render(): JSX.Element {
-        return (<span style={childStyle}>Hello UniteJS World!</span>);
+        return (<span className="child-style">Hello UniteJS World!</span>);
     }
 }