+++ assets/gulp/tasks/cssPre/less/build-css-components.js
@@ -19,7 +19,7 @@ gulp.task("build-css-components", async () => {
 
     let errorCount = 0;
 
-    return asyncUtil.stream(gulp.src(path.join(uniteConfig.dirs.www.src, "**/*.less"))
+    return asyncUtil.stream(gulp.src(path.join(uniteConfig.dirs.www.src, `**/*.${uniteConfig.styleExtension}`))
         .pipe(buildConfiguration.sourcemaps ? sourcemaps.init() : gutil.noop())
         .pipe(less())
         .on("error", (err) => {